Promote downloader to the domain
It obviously belongs there and we shouldn't get any funny ideas about
databases in its viccinity.
